third_party/blink/web_tests/external/wpt/html/canvas/element/text-styles/2d.text.draw.space.collapse.space-expected.txt

This is a testharness.js-based test.
FAIL Space characters are converted to U+0020, and collapsed (per CSS) assert_approx_equals: Red channel of the pixel at (25, 25) expected 0 +/- 2 but got 255
Harness: the test ran to completion.
